+++ lib/models/lists/list_model_animated.dart
