+++ lib/pages/select_timer/select_timer.dart
@@ -1,0 +1,63 @@
+import 'package:flutter/material.dart';
+import '../../utils/functions.dart';
+import '../../models/workout_type.dart';
+import 'widgets/timer_option_card.dart';
+
+class SelectTimer extends StatefulWidget {
+  const SelectTimer({super.key});
+
+  @override
+  SelectTimerState createState() => SelectTimerState();
+}
+
+class SelectTimerState extends State<SelectTimer> {
+  /// Since this will be a new timer, go ahead and create an
+  /// empty workout to pass to the next views.
+  ///
+  final workout = Workout.empty();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        appBar: AppBar(),
+        body: Padding(
+            padding: const EdgeInsets.fromLTRB(20, 5, 20, 5),
+            child: Column(
+              children: [
+                /// Card for the interval timer option.
+                ///
+                TimerOptionCard(
+                    onTap: () {
+                      pushCreateTimer(workout, context, false, (value) {});
+                    },
+                    optionIcon: Icons.timer,
+                    optionTitle: "Interval Timer",
+                    optionDescription:
+                        "An interval timer is a tool that helps you track the time spent working and resting during a workout."),
+
+                /// Card for the workout option.
+                ///
+                TimerOptionCard(
+                  onTap: () {
+                    pushCreateWorkout(workout, context, false, (value) {});
+                  },
+                  optionIcon: Icons.fitness_center,
+                  optionTitle: "Workout",
+                  optionDescription:
+                      "A workout is a planned set of exercises combined with an interval timer.",
+                ),
+
+                /// Card for the import option.
+                ///
+                TimerOptionCard(
+                  onTap: () {
+                    pushImportWorkout(workout, context, (value) {});
+                  },
+                  optionIcon: Icons.upload_file,
+                  optionTitle: "Import",
+                  optionDescription: "Import a workout or timer from file.",
+                ),
+              ],
+            )));
+  }
+}
